Fetch providers and owner refs to test patterns
This commit mixes restructuring with adding functionality because I'm still
figuring out the best approach here. The regenerating of partially generated
resolvers in place was a little magic, so I've disabled it. I also realised
that we need some concrete but generic type that can satisfy the
KubernetesResource interface in GraphQL, e.g. when resolving references in
contexts where Crossplane makes no guarantees about what class of resource the
reference will be to (i.e. when we don't know it will be a managed resource or
an XR).

Signed-off-by: Nic Cope <negz@rk0n.org>