Add debug logging, discover APIs at startup
Initially the first request made by each person was taking a huge ~17 seconds.
This turned out to be because the cache and the client were each taking ~8
seconds to perform API discovery. I added debug logging to diagnose this, and
moved REST mapping to startup. The first request made by each person in a small
cluster is now more or less instantaneous.

Signed-off-by: Nic Cope <negz@rk0n.org>