XRC resource references are optional
A composite resource claim may _technically_ exist with an unset reference to a
composite resource. Interestingly we probably want this to be a required field -
a claim without a resourceRef will sit around doing nothing until someone fixes
it. Unfortunately this is now part of the surface area of our v1 API and thus it
will remain optional.

Signed-off-by: Nic Cope <negz@rk0n.org>