Add an id for all distinct Kubernetes resources
The idea is to store the data that uniquely identifies a resource in the
Kubernetes API such that we can extract that data from a given ID
string in future. Kubernetes resources are uniquely identified by their API
version, kind, namespace (where relevant), and name. This set of data is often
called 'a reference' or 'an object reference'. While all of this data already
exists in our API representing it as a string gives GraphQL clients a single,
idiomatic scalar field they may consider the "primary key" of a type.

Each ID is a Kubernetes reference serialised as "apiVersion|kind|namespace|name"
and base64 encoded in a mild attempt to reinforce the fact that clients must
treat it as opaque. Cluster scoped resources have an empty namespace 'field',
i.e. "apiVersion|kind||name".

Signed-off-by: Nic Cope <negz@rk0n.org>