Use <T>Connection consistently
Prior to this commit we used <T>Connection to model connections between things
and <T>List to model query results. Both types are identical, so we're deduping
to Connection only. A query being 'connected' to stuff doesn't make a lot of
sense, but hey this pattern is good enough for GitHub so ¯\_(ツ)_/¯

Signed-off-by: Nic Cope <negz@rk0n.org>