Add tests for all implemented resolvers
The one case I'm frequently not testing is when the model conversion code fails.
In practice this code would only fail if passed a type that was not marshallable
as JSON. We might consider adding a MustMarshal helper that panics if Marshal
fails, since I believe this would only be possible if the program was broken (as
opposed to because we received invalid inputs).

Signed-off-by: Nic Cope <negz@rk0n.org>