Use a more idiomatic form of Connections
This commit disables the "relay-connection-types-spec" GraphQL schema linter,
which ensures that <T>Connection types comply with section 2 of the Relay
Connections spec.

Implementing the connection spec requires implementing pagination. We're not
doing that for now, but we want to leave room in our API to do so in future.
When we do add pagination we'd like it to comply with the connection spec, and
we'd like to add it without making a breaking change.

In the meantime, I've changed the 'items' and 'count' connection field to
'nodes' and 'totalCount', which appear to be more idiomatic. 'totalCount' is
mentioned in the GraphQL documentation, and 'nodes' is used by the GitHub API
alongside 'edges' as a simple array of nodes without cursor details. This will
allow us to add 'edges' and 'pageInfo' in future to become connection spec
compliant without a breaking change.

https://relay.dev/graphql/connections.htm#sec-Connection-Types

Signed-off-by: Nic Cope <negz@rk0n.org>