Support resolving events
This commit adds a generic Kubernetes event resolver that is used by all
existing implementations of the KubernetesResource GraphQL interface (i.e.
all existing resolvers that previously had stub 'events' methods). It also
implements the GenericResource and Secret resolvers, because they only resolve
events. A resolver _for_ (as opposed to of) events allows the involved object of
an event to be resolved.

Furthermore, this commit adds a new 'events' query that can be used to query for
events pertaining to a specific involved object by ID, or for all known events.

Signed-off-by: Nic Cope <negz@rk0n.org>