Remove specific 'list this type' queries with a kubernetesResources query
I felt uncomfortable with the implementation of the 'list this (class of) type'
queries like managedResources and compositeResources because in practice xgql
must make some surprising calls in order to answer the questions. For example to
get all managed resources of a particular kind given their CRD's ID we must
first get the CRD, which seems like something the caller may not expect and
which they may not have RBAC permission for.

In most cases I think it's slightly more intuitive to:

1. Use the kubernetesResource query to get a CRD or XRD by its ID
2. Use an inline fragment to address the returned interface as a CRD or XRD
3. Use its definedResources method to return defined resources.

This does require the caller to make a more complicated query, but that query
communicates to the caller what is going to happen; we're going to get a CRD or
XRD in order to look up the resources it defines.

Signed-off-by: Nic Cope <negz@rk0n.org>