Remove "return a one-item array" style query arguments
In cases where a particular argument would cause us to return a one item array
(like "get the active revision", or "get the controller reference") it would be
more convenient for the caller to use a distinct field like getController or
getActiveRevision. This commit removes the current implementation - the proposed
approach will be added in a future commit.

Signed-off-by: Nic Cope <negz@rk0n.org>