Add a StringMap scalar, and use it where we previously used Map
The Map scalar is a built in type of gqlgen that corresponds to a Go
map[string]interface{}. In practice we always want a map[string]string, so this
commit introduces a StringMap that corresponds to a map[string]string. This
allows callers to query for (e.g.) metadata labels or annotations and get a JSON
object of string key to string value in response. Note that this is still a
scalar, so it's not possible to query for only a particular key in the map.

Signed-off-by: Nic Cope <negz@rk0n.org>