Ensure connection arrays are stably sorted
Some of the node arrays returned by xgql are built from a Kubernetes client-go
cache. This cache is backed by a map, which returns elements in random order
when ranged over. This commit ensures we return such arrays sorted by ID (e.g.
GVK, namespace, and name). Note that connections which are not backed by a cache
list (e.g. connections that are populated by ranging over an ordered array of
references) are not sorted.

Signed-off-by: Nic Cope <negz@rk0n.org>