Perform API discovery more aggressively, but less frequently
This commit reduces the time it takes to perform API discovery on a kind cluster
with a moderate number of API endpoints (three providers with ~120 total CRDs
installed) from ~16 seconds to ~40ms. It does so by dramatically raising API
discovery rate limits, from 5qps with a 10qps burst to 20qps with a 100qps
burst.

To compensate for this, it also restricts discovery to happening no more than
once every 20 seconds, rather than up to 5 times per second. Discovery happens
once on startup, and then once any time a client requests a kind of API resource
that did not exist last time discovery was performed (e.g. because a provider
has installed new CRDs since). The implementation of the underlying Dynamic REST
mapper means that discovery results are now effectively cached for 20 seconds.

Signed-off-by: Nic Cope <negz@rk0n.org>