Improve the heuristics we use to identify resource types
This makes GetKubernetesResource a little more robust in the face of ambiguous
resources, for example an XR that has a composition selector set but that is not
able to set its composition ref because no composition matches.

These heuristics are still flawed; for example we would be unable to identify a
pathological XR with no compositionRef, compositionSelector, or resourceRefs.

Signed-off-by: Nic Cope <negz@rk0n.org>