Have 'make generate' add license headers to all files
Use https://github.com/google/addlicense to add an Apache license (the default)
to all Go files. This tool unfortunately uses // comments rather than /* */.
This is a little less pleasing to my eye, but worth it for the automation. Files
are automatically licensed as the current year (unless they already have a
license header).

Signed-off-by: Nic Cope <negz@rk0n.org>